# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/me": {
      "get": {
        "operationId": "CoreAuthController_me_v1",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUser_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Current user profile",
        "tags": ["auth"]
      },
      "patch": {
        "operationId": "CoreAuthController_updateProfile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileInput"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUser_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update the current user's editable profile fields",
        "tags": ["auth"]
      },
      "delete": {
        "operationId": "CoreAuthController_deleteMe_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Hard-delete the current user. Cascades sessions, refresh tokens, OTPs, and OAuth links. Audit events survive.",
        "tags": ["auth"]
      }
    },
    "/v1/auth/accounts/{provider}": {
      "delete": {
        "operationId": "CoreAuthController_unlinkOAuthAccount_v1",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Unlink an OAuth identity. Refuses if doing so would leave the user with no remaining auth method.",
        "tags": ["auth"]
      }
    },
...
    "/v1/auth/google": {
      "post": {
        "description": "Exchanges a Google ID token (from Google Identity Services on web, or the native Google Sign-In SDK on mobile) for a first-party session. Verification is local: we check the JWT signature, expiry, and audience against the configured `GOOGLE_CLIENT_ID_*` client IDs. Google's token is discarded after verification — subsequent requests authenticate with the access/refresh tokens this endpoint returns.",
        "operationId": "GoogleAuthController_signin_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInWithGoogleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider identity resolved; first-party session issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair_Output"
                }
              }
            }
          },
          "202": {
            "description": "The provider did not verify the email. A one-time code was sent and must be submitted to /v1/auth/oauth-email/verify.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpChallengeMetadata_Output"
                }
              }
            }
          },
          "401": {
            "description": "The Google identity token failed verification."
          }
        },
        "summary": "Sign in with a Google-issued ID token",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "Turborepo Full Template API",
    "description": "REST API surface. All operations live under /v1.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "HTTP-only access-token cookie."
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TokenPair_Output": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Signed access JWT. Also set as the `access_token` HttpOnly cookie; mobile clients consume the body field."
          },
          "refreshToken": {
            "type": "string",
            "description": "Signed refresh JWT. Also set as the `refresh_token` HttpOnly cookie."
          }
        },
        "required": ["accessToken", "refreshToken"],
        "additionalProperties": false
      },
...
      "SessionUser_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "string",
            "description": "ISO timestamp of email verification, or `null` if unverified.",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "phoneVerified": {
            "type": "string",
            "description": "ISO timestamp of phone verification, or `null` if unverified.",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Authorization roles. Snapshot also lives in the access-JWT claims; reflects this user's current grants."
          },
          "createdAt": {
            "type": "string",
            "description": "ISO timestamp of account creation."
          },
          "linkedProviders": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["google", "apple"]
            }
          }
        },
        "required": [
          "id",
          "email",
          "emailVerified",
          "phone",
          "phoneVerified",
          "firstName",
          "lastName",
          "roles",
          "createdAt",
          "linkedProviders"
        ],
        "additionalProperties": false
      },
      "UpdateProfileInput": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OtpChallengeMetadata_Output": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["verification_required"]
          },
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "Opaque UUID identifying one OTP challenge."
          },
          "expiresInSec": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "description": "Seconds until the challenge expires."
          },
          "resendAfterSec": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Seconds until another delivery may be requested."
          }
        },
        "required": ["status", "challengeId", "expiresInSec", "resendAfterSec"],
        "additionalProperties": false
      },
...
      "SignInWithGoogleInput": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "minLength": 20,
            "description": "Google-issued ID token (JWT) from Google Identity Services or the native SDK."
          }
        },
        "required": ["idToken"],
        "additionalProperties": false
      },
...
    }
  }
}